infra: relocate app runner and implement User-Agent generator using package version
