+++ internal/infra/app/app.go
@@ -1,4 +1,4 @@
-package otomo
+package app
 
 import (
 	"context"
