+++ internal/infra/tool/user_agent.go
@@ -1,0 +1,11 @@
+package tool
+
+import (
+	"fmt"
+
+	"github.com/handlename/otomo"
+)
+
+func UserAgent() string {
+	return fmt.Sprintf("otomo-bot/%s", otomo.Version)
+}
